remove unnecessary useHeaderHeight usage
