+++ apps/www/components/Questionnaire/Submissions/index.tsx
@@ -37,7 +37,6 @@ const Page = ({
   extract,
   share,
 }: SubmissionsOverviewProps) => {
-  const headerHeight = useHeaderHeight()
   const router = useRouter()
   const { query } = router
   const { questionIds, submissionId } = mapShareParam(query.share)
@@ -49,7 +48,7 @@ const Page = ({
     if (query?.focus === MAIN_VIEWPORT_FOCUS) {
       scrollIntoView(answersRef.current, {
         time: 0,
-        align: { topOffset: headerHeight, top: 0 },
+        align: { top: 0 },
       })
     }
   }, [])
