+++ apps/www/components/Questionnaire/Submissions/views/QuestionView.jsx
@@ -9,7 +9,6 @@ import {
   Loader,
   Editorial,
   Center,
-  useHeaderHeight,
 } from '@project-r/styleguide'
 import { css } from 'glamor'
 
@@ -41,7 +40,6 @@ const QuestionView = ({
   questionnaireBgColor,
 }) => {
   const { t } = useTranslation()
-  const headerHeight = useHeaderHeight()
   const { loading, error, data, fetchMore } = useQuery(
     QUESTIONNAIRE_WITH_SUBMISSIONS_QUERY,
     {
@@ -68,7 +66,7 @@ const QuestionView = ({
     if (extract) return
     scrollIntoView(questionRef.current, {
       time: 0,
-      align: { topOffset: headerHeight, top: 0 },
+      align: { top: 0 },
     })
   }, [])
