+++ apps/www/components/Questionnaire/Submissions/views/SubmissionView.tsx
@@ -12,7 +12,6 @@ import {
   Center,
   ColorContextProvider,
   Editorial,
-  useHeaderHeight,
 } from '@project-r/styleguide'
 
 import { OverviewLink } from '../components/Links'
@@ -31,7 +30,6 @@ const SubmissionView = ({
   const { t } = useTranslation()
   const router = useRouter()
   const pathname = router.asPath.split('?')[0]
-  const headerHeight = useHeaderHeight()
 
   const { loading, error, data } = useQuery(
     QUESTIONNAIRE_WITH_SUBMISSIONS_QUERY,
@@ -55,7 +53,7 @@ const SubmissionView = ({
     if (extract) return
     scrollIntoView(submissionRef.current, {
       time: 0,
-      align: { topOffset: headerHeight, top: 0 },
+      align: { top: 0 },
     })
   }, [])
