+++ packages/styleguide/src/lib/useHeaderHeight.docs.md
@@ -1,4 +1,4 @@
-Utility hook passing the height of the header to its children by way of a context payload. Additionally
+Utility hook passing the height of the header to its children by way of a context payload.
 
 Example using the `headerHeight` value programmatically:
 
@@ -14,7 +14,7 @@ const Child = () => {
   )
 }
 
-<HeaderHeightProvider config={[{ minWidth: 0, headerHeight: 30 }]}>
+<HeaderHeightProvider height={height: SOME_DYNAMIC_HEIGHT_VALUE}>
   <Child />
 </HeaderHeightProvider>
 ```
